% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{chomperCAVI}
\alias{chomperCAVI}
\title{CHOMPER with a single Coordinate Ascent Variational Inference}
\usage{
chomperCAVI(
  x,
  k,
  n,
  N,
  p,
  M,
  discrete_fields,
  continuous_fields,
  hyper_beta,
  hyper_phi = c(),
  hyper_tau = c(),
  hyper_epsilon_discrete = c(),
  hyper_epsilon_continuous = c(),
  hyper_sigma = matrix(nrow = 0, ncol = 2),
  overlap_prob = 0.5,
  tol_cavi = 1e-05,
  max_iter_cavi = 100,
  max_time = 86400,
  custom_initializer = FALSE,
  use_checkpoint = FALSE,
  initial_values = NULL,
  checkpoint_values = NULL,
  verbose_internal = TRUE
)
}
\arguments{
\item{x}{A list of data frames, each representing a dataset.}

\item{k}{The number of datasets to be linked.}

\item{n}{The number of rows in each dataset (vector of length k).}

\item{N}{The number of columns in each dataset.}

\item{p}{The number of fields in each dataset.}

\item{M}{The number of categories for each discrete field (vector of length of discrete fields).}

\item{discrete_fields}{The indexes of the discrete fields (1-based index).}

\item{continuous_fields}{The indexes of the continuous fields (1-based index).}

\item{hyper_beta}{The hyperparameters for the beta distribution (matrix of size p x 2).}

\item{hyper_phi}{The hyperparameters for softmax representation (vector of length of discrete fields).}

\item{hyper_tau}{The temperature parameter (vector of length of discrete fields).}

\item{hyper_epsilon_discrete}{The range parameter for the comprehensive hit of discrete fields (vector of length of discrete fields).}

\item{hyper_epsilon_continuous}{The range parameter for the comprehensive hit of continuous fields (vector of length of continuous fields).}

\item{hyper_sigma}{The hyperparameters for the Inverse Gamma distribution (matrix of size length of continuous fields x 2).}

\item{overlap_prob}{The presumed probability of overlap across the datasets.}

\item{tol_cavi}{The tolerance for the coordinate ascent variational inference for the convergence.}

\item{max_iter_cavi}{The maximum number of iterations for the coordinate ascent variational inference.}

\item{max_time}{The maximum time limit for the execution in seconds.}

\item{custom_initializer}{Whether to use a custom initializer for the initial values.}

\item{use_checkpoint}{Whether to use a checkpoint.}

\item{initial_values}{The initial values for the parameters (optional).}

\item{checkpoint_values}{The checkpoint values for the parameters (optional).}

\item{verbose_internal}{Whether to print the internal C++ messages (TRUE: print, FALSE: not print).}
}
\value{
A list of the approximated parameters of the variational factors and other information containing:
\itemize{
\item \code{nu}: A list of parameter matrices for the approximate multinomial posterior of the linkage structure.
\item \code{omega}: A matrix of parameter vectors for the approximate beta posterior of the distortion ratio.
\item \code{rho}: A list of parameter matrices for the approximate Bernoulli posterior of the distortion indicators.
\item \code{gamma}: A list of parameter matrices for the approximate multinomial posterior of discrete true latent values.
\item \code{alpha}: A list of parameter vectors for the approximate Dirichlet posterior (theta) of the discrete true latent values.
\item \code{eta_tilde}: A matrix of parameter vectors for the mean of the approximate normal posterior of continuous true latent values.
\item \code{eta_mean}: A vector of mean parameters for the approximate normal posterior (eta_tilde) of the continuous true latent values.
\item \code{eta_var}: A vector of variance parameters for the approximate normal posterior (eta_tilde) of the continuous true latent values.
\item \code{sigma_tilde}: A matrix of parameter vectors for the variance of the approximate normal posterior of continuous true latent values.
\item \code{sigma_shape}: A vector of shape parameters for the approximate inverse gamma posterior (sigma_tilde) of the continuous true latent values.
\item \code{sigma_scale}: A vector of scale parameters for the approximate inverse gamma posterior (sigma_tilde) of the continuous true latent values.
\item \code{ELBO}: The final maximum ELBO from a single CAVI.
\item \code{niter}: The number of iterations for a single CAVI.
\item \code{interruption}: Whether the CHOMPER-CAVI is interrupted. The fitting is interrupted if the elapsed time reaches the maximum time limit.
\item \code{cavi_elapsed_time}: The maximum elapsed time of a single CAVI iteration.
\item \code{elapsed_time}: The elapsed time of the entire CAVI process in seconds.
}
}
\description{
Fit the CHOMPER model with a single Coordinate Ascent Variational Inference (CAVI) to estimate the linkage structure across multiple datasets.
It returns the approximate variational factors of the linkage structure that maximize the evidence lower bound (ELBO) and other parameters of the CHOMPER model.
}
\examples{
# 1. Generate sample data for testing
sample_data <- generate_sample_data(
  n_entities = 10,
  n_files = 3,
  overlap_ratio = 0.7,
  discrete_columns = c(1, 2),
  discrete_levels = c(3, 3),
  continuous_columns = c(3, 4),
  continuous_params = matrix(c(0, 0, 1, 1), ncol = 2),
  distortion_ratio = c(0.1, 0.1, 0.1, 0.1)
)

# 2. Get file information and remove `id` from the original data
n <- numeric(3)
x <- list()
for (i in 1:3) {
  n[i] <- nrow(sample_data[[i]])
  x[[i]] <- sample_data[[i]][, -1]
}
N <- sum(n)

# 3. Set Hyperparameters
hyper_beta <- matrix(
  rep(c(N * 0.1 * 0.01, N * 0.1), 4),
  ncol = 2, byrow = TRUE
)

hyper_sigma <- matrix(
  rep(c(0.01, 0.01), 2),
  ncol = 2, byrow = TRUE
)

# 4. Fit CHOMPER-CAVI
result <- chomperCAVI(
  x = x,
  k = 3, # number of datasets
  n = n, # rows per dataset
  N = N, # columns per dataset
  p = 4, # fields per dataset
  M = c(3, 3), # categories for discrete fields
  discrete_fields = c(1, 2),
  continuous_fields = c(3, 4),
  hyper_beta = hyper_beta, # hyperparameter for distortion rate
  hyper_sigma = hyper_sigma, # hyperparameter for continuous fields
  hyper_phi = c(2.0, 2.0),
  hyper_tau = c(0.01, 0.01),
  hyper_epsilon_discrete = c(0, 0),
  hyper_epsilon_continuous = c(0.001, 0.001),
)

}
