% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{chomperDIG}
\alias{chomperDIG}
\title{CHOMPER with Discomfort-Informed Adaptive Gibbs Sampler}
\usage{
chomperDIG(
  x,
  k,
  n,
  N,
  p,
  M,
  discrete_fields,
  continuous_fields,
  hyper_beta,
  hyper_phi = c(),
  hyper_tau = c(),
  hyper_epsilon_discrete = c(),
  hyper_epsilon_continuous = c(),
  hyper_sigma = matrix(nrow = 0, ncol = 2),
  decaying_upper_bound = 100,
  n_burnin = 1000,
  n_gibbs = 1000,
  batch_size = 10,
  n_epochs = 5,
  max_time = 86400,
  batch_update = TRUE,
  verbose_internal = TRUE
)
}
\arguments{
\item{x}{A list of data frames, each representing a dataset.}

\item{k}{The number of datasets to be linked.}

\item{n}{The number of rows in each dataset (vector of length k).}

\item{N}{The number of columns in each dataset.}

\item{p}{The number of fields in each dataset.}

\item{M}{The number of categories for each discrete field (vector of length of discrete fields).}

\item{discrete_fields}{The indexes of the discrete fields (1-based index).}

\item{continuous_fields}{The indexes of the continuous fields (1-based index).}

\item{hyper_beta}{The hyperparameters for the beta distribution (matrix of size p x 2).}

\item{hyper_phi}{The hyperparameters for softmax representation (vector of length of discrete fields).}

\item{hyper_tau}{The temperature parameter (vector of length of discrete fields).}

\item{hyper_epsilon_discrete}{The range parameter for the comprehensive hit of discrete fields (vector of length of discrete fields).}

\item{hyper_epsilon_continuous}{The range parameter for the comprehensive hit of continuous fields (vector of length of continuous fields).}

\item{hyper_sigma}{The hyperparameters for the Inverse Gamma distribution (matrix of size length of continuous fields x 2).}

\item{decaying_upper_bound}{The upper bound for the decaying parameter for the discomfort-informed adaptive Gibbs sampler.}

\item{n_burnin}{The number of burn-in iterations for the MCMC.}

\item{n_gibbs}{The number of Gibbs sampling iterations for the MCMC.}

\item{batch_size}{The batch size for the discomfort-informed adaptive Gibbs sampler.}

\item{n_epochs}{The number of epochs for the discomfort-informed adaptive Gibbs sampler.}

\item{max_time}{The maximum time limit for the execution in seconds.}

\item{batch_update}{Whether to update the parameters in a batch.}

\item{verbose_internal}{Whether to print the internal C++ messages (TRUE: print, FALSE: not print).}
}
\value{
A list containing the posterior samples.

A list of the posterior samples and other information containing:
\itemize{
\item \code{lambda}: A list of posterior samples (integer vectors) of the linkage structure.
\item \code{z}: A list of posterior samples (binary matrices) of the distortion indicators.
\item \code{y}: A list of posterior samples (matrices) of the true latent records.
\item \code{beta}: A list of posterior samples (numeric vectors) of the distortion ratio.
\item \code{theta}: A list of posterior samples (numeric vectors) of the probabilities of discrete true latent values.
\item \code{eta}: A list of posterior samples (numeric vectors) of the mean of continuous true latent values.
\item \code{sigma}: A list of posterior samples (numeric vectors) of the variance of continuous true latent values.
\item \code{n_sample}: Total number of posterior samples after burn-in.
\item \code{elapsed_time}: The elapsed time of the entire MCMC process in seconds.
\item \code{interruption}: Whether the CHOMPER-DIG is interrupted. The fitting is interrupted if the elapsed time reaches the maximum time limit.
}
}
\description{
Fit the CHOMPER model with Discomfort-Informed Adaptive Gibbs Sampler (DIG) to estimate the linkage structure across multiple datasets.
It returns the posterior samples of the linkage structure and other parameters of the CHOMPER model.
}
\examples{
# 1. Generate sample data for testing
sample_data <- generate_sample_data(
  n_entities = 10,
  n_files = 3,
  overlap_ratio = 0.7,
  discrete_columns = c(1, 2),
  discrete_levels = c(3, 3),
  continuous_columns = c(3, 4),
  continuous_params = matrix(c(0, 0, 1, 1), ncol = 2),
  distortion_ratio = c(0.1, 0.1, 0.1, 0.1)
)

# 2. Get file information and remove `id` from the original data
n <- numeric(3)
x <- list()
for (i in 1:3) {
  n[i] <- nrow(sample_data[[i]])
  x[[i]] <- sample_data[[i]][, -1]
}
N <- sum(n)

# 3. Set Hyperparameters
hyper_beta <- matrix(
  rep(c(N * 0.1 * 0.01, N * 0.1), 4),
  ncol = 2, byrow = TRUE
)

hyper_sigma <- matrix(
  rep(c(0.01, 0.01), 2),
  ncol = 2, byrow = TRUE
)

# 4. Fit CHOMPER-DIG
result <- chomperDIG(
  x = x,
  k = 3, # number of datasets
  n = n, # rows per dataset
  N = N, # columns per dataset
  p = 4, # fields per dataset
  M = c(3, 3), # categories for discrete fields
  discrete_fields = c(1, 2),
  continuous_fields = c(3, 4),
  hyper_beta = hyper_beta, # hyperparameter for distortion rate
  hyper_sigma = hyper_sigma, # hyperparameter for continuous fields
  hyper_phi = c(2.0, 2.0),
  hyper_tau = c(0.01, 0.01),
  hyper_epsilon_discrete = c(0, 0),
  hyper_epsilon_continuous = c(0.001, 0.001),
  decaying_upper_bound = 100.0,
  n_burnin = 0,
  n_gibbs = 100,
  batch_size = 5,
  n_epochs = 5
)

}
