% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{flatten_posterior_samples}
\alias{flatten_posterior_samples}
\title{Flatten the posterior samples, lambda, into a matrix}
\usage{
flatten_posterior_samples(samples, k, N)
}
\arguments{
\item{samples}{a list of MCMC samples}

\item{k}{number of files to be linked}

\item{N}{total number of records}
}
\value{
an N by number of MCMC samples matrix
}
\description{
This function converts a list of posterior samples of lambda into a matrix.
Before calculating the posterior similarity matrix using \code{psm_mcmc}, it is necessary to flatten the posterior samples into a matrix.
}
\examples{
# 1. Create a list of posterior samples of linkage structure
number_of_files <- 2

n_file1 <- 2
n_file2 <- 3

number_of_records <- n_file1 + n_file2

number_of_samples <- 10
lambda <- list()
for (i in 1:number_of_samples) {
  lambda[[i]] <- list(
    sample(1:number_of_records, n_file1, TRUE),
    sample(1:number_of_records, n_file2, TRUE)
  )
}

# 2. Converts a list of posterior samples of lambda into a matrix
flatten_posterior_samples(lambda, number_of_files, number_of_records)

}
