% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_data.R
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Generate synthetic data for record linkage}
\usage{
generate_sample_data(
  n_entities,
  n_files,
  overlap_ratio,
  discrete_columns,
  discrete_levels,
  continuous_columns,
  continuous_params,
  distortion_ratio,
  discrete_fuzziness = NULL,
  continuous_fuzziness = NULL
)
}
\arguments{
\item{n_entities}{The number of entities.}

\item{n_files}{The number of files.}

\item{overlap_ratio}{The ratio of overlaping entities across the files.}

\item{discrete_columns}{The indices of the discrete columns (1-based index).}

\item{discrete_levels}{The levels of the discrete columns (vector of length of discrete columns).}

\item{continuous_columns}{The indices of the continuous columns (1-based index).}

\item{continuous_params}{The parameters of the continuous columns (matrix of size length of continuous columns x 2).}

\item{distortion_ratio}{The distortion ratio of the columns (vector of length of total columns).}

\item{discrete_fuzziness}{The configuration of the multiple truths of the discrete columns (optional).}

\item{continuous_fuzziness}{The configuration of the multiple truths of the continuous columns (optional).}
}
\value{
A list of matrices containing the noisy synthetic data. Each matrix represents a file.
}
\description{
Generate synthetic data for record linkage with given number of entities, files, and overlap ratio.
Each variable can follow either a multinomial or a Gaussian distribution.
User can specify the existence of multiple truths for each variable.
}
\examples{
# 1. Set number of entities, files, and overlap ratio
n_entities <- 25
n_files <- 2
overlap_ratio <- 0.9

# 2. Set attributes information
discrete_columns <- 1:4
discrete_levels <- rep(5, 4)
continuous_columns <- 5:6
continuous_params <-
  matrix(c(0, 10, 10, 10),
    ncol = 2, byrow = TRUE # means and variances
  )

# 3. Set distortion ratio and fuzziness information
distortion_ratio <- rep(0.01, 6)
discrete_fuzziness <- matrix(c(4, 1),
  ncol = 2, byrow = TRUE
)
continuous_fuzziness <- matrix(c(5, 0.5^2, 6, 0.5^2),
  ncol = 2, byrow = TRUE
)

# 4. Generate synthetic data
simulation_data <- generate_sample_data(
  n_entities, n_files, overlap_ratio,
  discrete_columns, discrete_levels,
  continuous_columns, continuous_params,
  distortion_ratio,
  discrete_fuzziness, continuous_fuzziness
)

}
