% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy}
\alias{italy}
\title{Italian Survey on Household Income and Wealth (ISHIW) data from 2020 and 2022}
\format{
A list of matrices with 9 variables. 15,198 rows and 23,057 rows are contained in surveys in 2020 and 2022, respectively.
\describe{
\item{ID}{Unique ID of each entity}
\item{SEX}{sex}
\item{ANASC}{year of birth}
\item{CIT}{nationality (whether or not Italian)}
\item{NASCREG}{administrative region of birth}
\item{STUDIO}{education level}
\item{IREG}{administrative region of current residence}
\item{NETINCOME}{net income}
\item{VALABIT}{estimated price of residence}
}
}
\source{
\url{https://www.bancaditalia.it/statistiche/tematiche/indagini-famiglie-imprese/bilanci-famiglie/distribuzione-microdati/index.html}
}
\usage{
italy
}
\description{
Two data files collected in 2020 and 2022 surveys.
Totally 38,255 records from 28,000 participants exist across two survey.
Categorical variables are recorded with integers, and continuous variables are standardized to have a mean of 0 and a variance of 1.
}
\keyword{datasets}
