% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{performance}
\alias{performance}
\title{Evaluate the performance of the linkage structure estimation}
\usage{
performance(estimation, truth, N, return_matrix = FALSE)
}
\arguments{
\item{estimation}{estimated linkage structure}

\item{truth}{true linkage structure}

\item{N}{total number of records}

\item{return_matrix}{if true, it also returns the matrix of linkage structure}
}
\value{
a list with performance metrics. If \code{return_matrix} is true, it also returns the matrix of linkage structure used for the evaluation.
}
\description{
Based on the true linkage structure and the estimate, it will calculate several metrics including true positive, true negative, false positive, false negative, false positive rate, and false negative rate.
This package recommends using the output of \code{links} function of the \code{blink} package as an argument to \code{performance} function.
}
\examples{
# 1. True linkage structure
total_number_of_records <- 6

truth <- matrix(
  list(c(1), c(2, 4), c(3), c(2, 4), c(5, 6), c(5, 6))
)

# 2. Estimated linkage structure
estimation <- matrix(
  list(c(1), c(2, 4), c(3), c(2, 4), c(5), c(6))
)

# 3. Calculate performance metrics
performance(estimation, truth, total_number_of_records, FALSE)

}
