% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{psm_mcmc}
\alias{psm_mcmc}
\title{Calculate the posterior similarity matrix}
\usage{
psm_mcmc(samples)
}
\arguments{
\item{samples}{a total number of records by number of MCMC samples matrix with MCMC samples}
}
\value{
a posterior similarity matrix of all possible pairs
}
\description{
This function returns a posterior similarity matrix based on the MCMC samples of lambda, obtained from \code{chomperMCMC}.
}
\examples{
# 1. Create a matrix with posterior samples of linkage structure
n_file1 <- 2
n_file2 <- 3

number_of_records <- n_file1 + n_file2

number_of_samples <- 10
lambda_matrix <- matrix(nrow = number_of_samples, ncol = number_of_records)
for (i in 1:number_of_samples) {
  lambda_matrix[i, ] <- sample(1:number_of_records, number_of_records, TRUE)
}

# 2. Calculate a posterior similarity matrix
psm_mcmc(lambda_matrix)

}
