% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{psm_vi}
\alias{psm_vi}
\title{Calculate the posterior similarity matrix}
\usage{
psm_vi(probs_field)
}
\arguments{
\item{probs_field}{a list of matrices with posterior probabilities}
}
\value{
a posterior similarity matrix of all possible pairs
}
\description{
This function returns a posterior similarity matrix based on the parameters of the approximated variational factor, nu, obtained from either \code{chomperEVIL} or \code{chomperCAVI}.
}
\examples{
# 1. Create an approximate posterior distribution of linkage structure
n_file1 <- 2
n_file2 <- 3

nu1 <- matrix(runif(n_file1^2 * n_file2), nrow = n_file1)
for (i in 1:n_file1) {
  nu1[i, ] <- nu1[i, ] / sum(nu1[i, ])
}

nu2 <- matrix(runif(n_file1 * n_file2^2), nrow = n_file2)
for (i in 1:n_file2) {
  nu2[i, ] <- nu2[i, ] / sum(nu2[i, ])
}

# 2. Convert into the appropriate type to run a function
approximate_posterior <- list(nu1, nu2)

# 3. Calculate a posterior similarity matrix
psm_vi(approximate_posterior)

}
