% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation.low}
\alias{simulation.low}
\title{Synthetic data with low overlap ratio (30\%)}
\format{
A list of matrices with 8 variables. 478 rows and 497 rows are contained in data, respectively.
\describe{
\item{id}{Unique ID of each entity}
\item{X01}{categorical variable with a single truth of 8 levels}
\item{X02}{categorical variable with a single truth of 8 levels}
\item{X03}{categorical variable with a single truth of 8 levels}
\item{X04}{categorical variable with 1-adjacent multiple truths of 8 levels}
\item{X05}{categorical variable with 1-adjacent multiple truths of 8 levels}
\item{X06}{continuous variable with multiple truths}
\item{X07}{continuous variable with multiple truths}
}
}
\source{
Generated by the package authors. Authors used \code{generate_sample_data} function in \code{R/generate_sample_data.R}.
}
\usage{
simulation.low
}
\description{
A synthetic data to illustrate the functionality of \code{chomper} model.
A list contains 2 matrices to perform entity resolution, and records are generated with 750 unique entities.
Each file contains 8 variables including unique id and 478 and 497 records, respectively.
Categorical variables are recorded with integers, and continuous variables are standardized to have a mean of 0 and a variance of 1.
}
\keyword{datasets}
