% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chomper-package.R
\docType{package}
\name{chomper-package}
\alias{chomper}
\alias{chomper-package}
\title{chomper: A Comprehensive Hit or Miss Probabilistic Entity Resolution Model}
\description{
Provides Bayesian probabilistic methods for record linkage and entity resolution across multiple datasets using the Comprehensive Hit Or Miss Probabilistic Entity Resolution (CHOMPER) model. The package implements four main inference approaches: (1) Evolutionary Variational Inference for record Linkage (EVIL), (2) Coordinate Ascent Variational Inference (CAVI), (3) Markov Chain Monte Carlo (MCMC) with split and merge process, and (4) Discomfort-Informed Adaptive Gibbs Sampler. The model supports both discrete and continuous fields, and it performs locally-varying hit mechanism for the attributes with multiple truths. It also provides tools for performance evaluation based on either approximated variational factors or posterior samples. The package is designed to support parallel computing with multi-threading support for EVIL to estimate the linkage structure faster.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hjkim8987/chomper}
  \item Report bugs at \url{https://github.com/hjkim8987/chomper/issues}
}

}
\author{
\strong{Maintainer}: Hyungjoon Kim \email{hjkim8987@gmail.com}

Authors:
\itemize{
  \item Andee Kaplan \email{Andee.Kaplan@colostate.edu}
  \item Matthew Koslovsky \email{Matt.Koslovsky@colostate.edu}
}

}
